Move disconnect module implementation to separate files.
