+++ tedge/src/cli/mod.rs
@@ -32,7 +32,7 @@ pub enum TEdgeOpt {
     Connect(connect::TEdgeConnectOpt),
 
     /// Remove bridge connection for a provider
-    Disconnect(disconnect::TedgeDisconnectBridgeOpt),
+    Disconnect(disconnect::TedgeDisconnectBridgeCli),
 
     /// Publish a message on a topic and subscribe a topic.
     Mqtt(mqtt::MqttCmd),
